create refactor of folder sturcture in some components
