+++ src/components/split-button/split-button.stories.js
@@ -4,8 +4,7 @@ import { text, boolean, select } from '@storybook/addon-knobs';
 import SplitButton from './split-button';
 import OptionsHelper from '../../utils/helpers/options-helper';
 import Button from '../button';
-import notes from './notes.md';
-import { StoryHeader, StoryCode, StoryCodeBlock } from '../../../.storybook/style/storybook-info.styles';
+import { notes, info } from './documentation';
 
 storiesOf('Split Button', module)
   .addParameters({
@@ -35,25 +34,7 @@ storiesOf('Split Button', module)
       );
     },
     {
-      info: {
-        text: (
-          <div>
-            <p> Split Button component </p>
-            <StoryHeader> Implementation</StoryHeader>
-
-            <p>Import the component:</p>
-            <StoryCode padded>import SplitButton from {'"react-carbon/lib/components/split-button"'}</StoryCode>
-
-            <p>To render a SplitButton (developer can add any buttons to dropdown):</p>
-            <StoryCodeBlock>
-              {'<SplitButton text="Main Button" onClick={ clickHandler }>'}
-              {'  <Button onClick="buttonClickHandler1">Button name 1</Button>'}
-              {'  <Button onClick="buttonClickHandler2">Button name 2</Button>'}
-              {'</SplitButton>'}
-            </StoryCodeBlock>
-          </div>
-        )
-      }
+      info: { text: info },
+      notes: { markdown: notes }
     },
-    { notes: { markdown: notes } }
   );
