+++ src/components/switch/documentation/notes.md
