+++ src/components/table-ajax/table-ajax.stories.js
@@ -3,13 +3,12 @@ import { storiesOf } from '@storybook/react';
 import { text, boolean } from '@storybook/addon-knobs';
 import { State, Store } from '@sambego/storybook-state';
 import { enableMock } from '../../../demo/xhr-mock';
-import notes from './notes.md';
 import {
   TableAjax, TableRow, TableCell, TableHeader
 } from './table-ajax';
 import Button from '../button';
 import MultiActionButton from '../multi-action-button';
-import { StoryHeader, StoryCode, StoryCodeBlock } from '../../../.storybook/style/storybook-info.styles';
+import { info, notes } from './documentation';
 
 const store = new Store({
   sortOrder: 'asc',
@@ -96,32 +95,7 @@ storiesOf('Table Ajax', module)
       );
     },
     {
-      info: {
-        text: (
-          <div>
-            <p> Table Ajax component</p>
-
-            <StoryHeader> Implementation</StoryHeader>
-            <StoryCodeBlock>
-              {'import Table from "carbon-react/lib/components/table-ajax"'}
-              {'import { TableRow, TableCell, TableHeader } from "carbon-react/lib/components/table"'}
-            </StoryCodeBlock>
-
-            <p>
-              To render a<StoryCode padded> {'Table'} </StoryCode>
-              please see the
-              <StoryCode padded> {'Table'} </StoryCode>
-              component
-            </p>
-
-            <p>
-              <StoryCode padded> {'Table'} </StoryCode>
-              requires a<StoryCode padded> {'path'} </StoryCode>
-              to be provided
-            </p>
-          </div>
-        )
-      }
+      info: { text: info },
+      notes: { markdown: notes }
     },
-    { notes: { markdown: notes } }
   );
