+++ src/components/textarea/documentation/index.js
@@ -1,0 +1,2 @@
+export { default as info } from './info';
+export { default as notes } from './notes.md';
